review follow up and add'l API documentation:
* made k_led_on/off/at public API functions
* removed unnecessary board specific impls of k_led_init
* added a missing header guard around led.h
