+++ hal/kubos-hal/source/led.c
@@ -19,11 +19,44 @@
 
 #include "kubos-hal/led.h"
 
+void k_led_init(int led_pin)
+{
+    k_gpio_init(led_pin, K_GPIO_OUTPUT, K_GPIO_PULL_NONE);
+}
+
 void k_led_init_all(void)
 {
     for (int i = 0; i < K_LED_COUNT; i++) {
-        k_led_init(K_LED_AT(i));
+        k_led_init(k_led_at(i));
     }
+}
+
+void k_led_on(int pin)
+{
+    k_gpio_write(pin, 1);
+}
+
+void k_led_off(int pin)
+{
+    k_gpio_write(pin, 0);
+}
+
+int k_led_at(int i)
+{
+    if (i < 0 || i >= K_LED_COUNT) return K_LED_UNKNOWN;
+
+#ifdef K_LED_0
+    if (i == 0) return K_LED_0;
+#endif
+#ifdef K_LED_1
+    if (i == 1) return K_LED_1;
+#endif
+#ifdef K_LED_2
+    if (i == 2) return K_LED_2;
+#endif
+#ifdef K_LED_3
+    if (i == 3) return K_LED_3;
+#endif
 }
 
 const char * k_led_get_desc(int led_pin)
